refactor: Migrate `provisioner.js` to TS
